+++ arcs/calc_iaa.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+python calc_iaa.py data/20150806/all.csv
