+++ arcs/data/README.md
@@ -1,0 +1,18 @@
+## Where did these files come from?
+
+If you go to the "RESULTS" section of your job, you can download results.
+
+Aggregated reports
+------------------------------
+* `all.csv` file was generated by going to "Settings" and selecting "All answers" in the dropdown.
+* `confidence_agg.csv` file was generated by selecting "Confidence Aggregation" with a Minimum Confidence of 0.0.
+
+
+Other reports
+------------------------------
+* The `annotator_info.csv` is the "Contributor" report.
+
+
+Notes
+------------------------------
+"Exclude test questions" was turned off when generating these results (ie. we *are* getting test questions in here)
